requirements.txt
requests
Pillow
pyheic